fixing Abstract Enemies
