+++ ForceGamaAttack/src/Observer/Observer.java
@@ -1,6 +1,8 @@
 package Observer;
+import entities.AbstractEnemy;
 import entities.AbstractEnemyFactory;
+import entities.Boss;
 
 public abstract class Observer {
-	public abstract void notifier(AbstractEnemyFactory enemy);
+	public abstract void notifier(AbstractEnemy boss);
 }
