+++ ForceGamaAttack/src/scenes/game/GameScene.java
@@ -1,5 +1,6 @@
 package scenes.game;
 import scenes.Scene;
+import entities.AbstractEnemy;
 import entities.AbstractEnemyFactory;
 import entities.EnemyType;
 import jplay.GameImage;
@@ -10,7 +11,7 @@ import jplay.Keyboard;
 public class GameScene extends Scene {
 	private GameImage background;
 	private GameImage playerImage;
-	private AbstractEnemyFactory enemy;
+	private AbstractEnemy enemy;
 	
 	protected void initialSetup(){
 		keyboard.setBehavior(Keyboard.DOWN_KEY, Keyboard.DETECT_EVERY_PRESS);
